+++ source/Internal/Transition/Adapter/TemplateLogic/FormatDateLogic.php
@@ -31,15 +31,15 @@ class FormatDateLogic
 
         // if such format applies to this type of field - sets formatted value to passed object
         if (!$myConfig->getConfigParam('blSkipFormatConversion')) {
-            if ($oConvObject->fldtype == "datetime" || $sFieldType == "datetime") {
+            if (($oConvObject && $oConvObject->fldtype == "datetime") || $sFieldType == "datetime") {
                 \OxidEsales\Eshop\Core\Registry::getUtilsDate()->convertDBDateTime($oConvObject);
-            } elseif ($oConvObject->fldtype == "timestamp" || $sFieldType == "timestamp") {
+            } elseif (($oConvObject && $oConvObject->fldtype == "timestamp") || $sFieldType == "timestamp") {
                 \OxidEsales\Eshop\Core\Registry::getUtilsDate()->convertDBTimestamp($oConvObject);
-            } elseif ($oConvObject->fldtype == "date" || $sFieldType == "date") {
+            } elseif (($oConvObject && $oConvObject->fldtype == "date") || $sFieldType == "date") {
                 \OxidEsales\Eshop\Core\Registry::getUtilsDate()->convertDBDate($oConvObject);
             }
         }
 
-        return $oConvObject->value;
+        return $oConvObject->value ?? '';
     }
 }